workload/requirements.txt
alembic~=1.13.3
asyncpg==0.29.0
dependency_injector==4.41.0
fastapi~=0.111.0
loguru==0.7.2
openpyxl==3.1.5
pandas==2.2.3
pydantic~=2.7.4
PyJWT~=2.8.0
python-dotenv==1.0.1
PyYAML~=6.0.1
SQLAlchemy~=2.0.31
uvicorn~=0.30.1
xlrd==2.0.1
python-multipart
sqladmin~=0.20.1
Flask~=3.0.3
psycopg2
Flask-SQLAlchemy
Flask-Admin
psycopg2
strawberry-graphql